remove duplicates and put every macro in EcalTools/macros
